Struct dxgcap::UniqueCOMPtr

pub struct UniqueCOMPtr<T: IUnknownT> {
    // some fields omitted
}

A unique pointer to a COM object. Handles refcounting.

Similar to windows CComQIPtr

Methods

impl<T: IUnknownT> UniqueCOMPtr<T>

unsafe fn new(ptr: *mut T) -> UniqueCOMPtr<T>

Construct a new unique COM pointer from a pointer to a COM interface. It is the users responsibility to guarantee that no copies of the pointer exists beforehand

unsafe fn query_interface<U>(self, interface_identifier: &IID) -> Result<UniqueCOMPtr<U>, HRESULT> where U: IUnknownT

Convert target interface by retrieving pointer to a supported interface of the object.

Examples

let output = {
    let mut output_ptr = ptr::null_mut();
    adapter.EnumOutputs(i, &mut output_ptr);
    UniqueCOMPtr::new(output)
};
let output1: UniqueCOMPtr<IDXGIOutput1> = output.query_interface(&IID_IDXGIOutput1).unwrap();

Safety

This method is unsafe, as coerced type may differ from the one requested by IID in query_interface. At the moment, following is legal:

let factory1: UniqueCOMPtr<IDXGIFactory1> = output.query_interface(&IID_IDXGIOutput1).unwrap();